Add files for D3DCompiler and dia2. Add support for dumping dia objects. Add support for reading raw ILDB parts and indexing tables by number.
