+++ tools/clang/tools/dotnetc/EditorForm.cs
@@ -43,6 +43,8 @@ namespace MainNs
         private List<DiagnosticDetail> diagnosticDetails;
         private DataGridView diagnosticDetailsGrid;
         private List<PassInfo> passInfos;
+        private TabPage debugInfoTabPage;
+        private RichTextBox debugInfoControl;
         private HlslHost hlslHost = new HlslHost();
 
         internal enum DocumentKind
@@ -510,11 +512,13 @@ namespace MainNs
                         }
                         else
                         {
+                            this.colorizationService.ClearColorization(this.DisassemblyTextBox);
                             this.DisassemblyTextBox.Text = GetStringFromBlob(result.GetErrors());
                         }
                     }
                     else
                     {
+                        this.colorizationService.ClearColorization(this.DisassemblyTextBox);
                         IDxcBlob code, errorMsgs;
                         uint flags = 0;
                         const uint D3DCOMPILE_DEBUG = 1;
@@ -688,16 +692,22 @@ namespace MainNs
         {
             private Dictionary<RichTextBox, RtbColorization> instances = new Dictionary<RichTextBox, RtbColorization>();
 
+            public void ClearColorization(RichTextBox rtb)
+            {
+                SetColorization(rtb, null);
+            }
+
             public void SetColorization(RichTextBox rtb, RtbColorization colorization)
             {
                 RtbColorization existing;
-                if (instances.TryGetValue(rtb, out existing))
+                if (instances.TryGetValue(rtb, out existing) && existing != null)
                 {
                     existing.Stop();
                 }
 
                 instances[rtb] = colorization;
-                colorization.Start();
+                if (colorization != null)
+                    colorization.Start();
             }
         }
 
@@ -3000,6 +3010,27 @@ namespace MainNs
             AnalysisTabControl.SelectTab(RewriterOutputTabPage);
         }
 
+        private IEnumerable<string> EnumerateDiaCandidates()
+        {
+            string progPath = Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86);
+            yield return System.IO.Path.Combine(progPath, @"Microsoft Visual Studio\2017\Community\DIA SDK\bin\amd64\msdia140.dll");
+            yield return System.IO.Path.Combine(progPath, @"Microsoft Visual Studio\2017\Community\Common7\IDE\msdia140.dll");
+            yield return System.IO.Path.Combine(progPath, @"Microsoft Visual Studio\2017\Community\Common7\IDE\msdia120.dll");
+        }
+
+        private string SuggestDiaRegistration()
+        {
+            foreach (string path in EnumerateDiaCandidates())
+            {
+                if (System.IO.File.Exists(path))
+                {
+                    return "Consider registering with this command:\r\n" +
+                        "regsvr32 \"" + path + "\"";
+                }
+            }
+            return null;
+        }
+
         private void debugInformationToolStripMenuItem_Click(object sender, EventArgs e)
         {
             if (this.SelectedShaderBlob == null)
@@ -3008,6 +3039,7 @@ namespace MainNs
                 return;
             }
 
+            this.ClearDiaUnimplementedFlags();
             IDxcBlob debugInfoBlob;
             dia2.IDiaDataSource dataSource;
             uint dfcc;
@@ -3020,7 +3052,26 @@ namespace MainNs
             }
             else
             {
-                dataSource = new dia2.DiaDataSource() as dia2.IDiaDataSource;
+                try
+                {
+                    dataSource = new dia2.DiaDataSource() as dia2.IDiaDataSource;
+                }
+                catch (System.Runtime.InteropServices.COMException ce)
+                {
+                    const int REGDB_E_CLASSNOTREG = -2147221164;
+                    if (ce.HResult == REGDB_E_CLASSNOTREG)
+                    {
+                        string suggestion = SuggestDiaRegistration();
+                        string message = "Unable to create dia object.";
+                        if (suggestion != null) message += "\r\n" + suggestion;
+                        MessageBox.Show(this, message);
+                    }
+                    else
+                    {
+                        HandleException(ce);
+                    }
+                    return;
+                }
                 dfcc = DFCC_SPDB;
             }
             uint index;
@@ -3049,14 +3100,202 @@ namespace MainNs
             var session = dataSource.openSession();
             var tables = session.getEnumTables();
             uint count = tables.get_Count();
-            string output = "Tables:\n";
+            StringBuilder output = new StringBuilder();
+            output.AppendLine("* Tables");
+            bool isFirstTable = true;
             for (uint i = 0; i < count; ++i)
             {
                 var table = tables.Item(i);
-                output += table.get_name();
-                output += "\n";
+                if (isFirstTable) isFirstTable = false; else output.AppendLine();
+                output.AppendLine("** " + table.get_name());
+                int itemCount = table.get_Count();
+                output.AppendLine("Record count: " + itemCount);
+                for (int itemIdx = 0; itemIdx < itemCount; itemIdx++)
+                {
+                    object o = table.Item((uint)itemIdx);
+                    if (TryDumpDiaObject(o as dia2.IDiaSymbol, output)) continue;
+                    if (TryDumpDiaObject(o as dia2.IDiaSourceFile, output)) continue;
+                    if (TryDumpDiaObject(o as dia2.IDiaLineNumber, output)) continue;
+                    if (TryDumpDiaObject(o as dia2.IDiaSectionContrib, output)) continue;
+                    if (TryDumpDiaObject(o as dia2.IDiaSegment, output)) continue;
+                }
             }
-            this.DisassemblyTextBox.Text = output;
+
+            if (debugInfoTabPage == null)
+            {
+                this.debugInfoTabPage = new TabPage("Debug Info");
+                this.debugInfoControl = new RichTextBox()
+                {
+                    Dock = DockStyle.Fill,
+                    Font = this.CodeBox.Font,
+                    ReadOnly = true
+                };
+                this.AnalysisTabControl.TabPages.Add(this.debugInfoTabPage);
+                this.debugInfoTabPage.Controls.Add(this.debugInfoControl);
+            }
+            this.debugInfoControl.Text = output.ToString();
+            this.AnalysisTabControl.SelectedTab = this.debugInfoTabPage;
+        }
+
+        private bool TryDumpDiaObject<TIface>(TIface o, StringBuilder sb)
+        {
+            if (o == null) return false;
+            DumpDiaObject(o, sb);
+            return true;
+        }
+
+        private void ClearDiaUnimplementedFlags()
+        {
+            foreach (var item in TypeWriters)
+                foreach (var writer in item.Value)
+                    writer.Unimplemented = false;
+        }
+
+        private void DumpDiaObject<TIface>(TIface o, StringBuilder sb)
+        {
+            Type type = typeof(TIface);
+            bool hasLine = false;
+            List<DiaTypePropertyWriter> writers;
+            if (SymTagEnumValues == null)
+            {
+                SymTagEnumValues = Enum.GetNames(typeof(dia2.SymTagEnum));
+            }
+            if (!TypeWriters.TryGetValue(type, out writers))
+            {
+                writers = new List<DiaTypePropertyWriter>();
+                foreach (System.Reflection.MethodInfo mi in type.GetMethods())
+                {
+                    Func<string, object, StringBuilder, bool> writer;
+                    if (mi.GetParameters().Length > 0)
+                        continue;
+                    string propertyName = mi.Name;
+                    if (!propertyName.StartsWith("get_")) continue;
+                    propertyName = propertyName.Substring(4);
+                    Type returnType = mi.ReturnType;
+                    if (returnType == typeof(string))
+                        writer = WriteDiaValueString;
+                    else if (returnType == typeof(uint) && propertyName == "symTag")
+                        writer = WriteDiaValueSymTag;
+                    else if (returnType == typeof(uint))
+                        writer = WriteDiaValueUInt32;
+                    else if (returnType == typeof(UInt64))
+                        writer = WriteDiaValueUInt64;
+                    else if (returnType == typeof(bool))
+                        writer = WriteDiaValueBool;
+                    else
+                        writer = WriteDiaValueAny;
+                    writers.Add(new DiaTypePropertyWriter()
+                    {
+                        MI = mi,
+                        PropertyName = propertyName,
+                        Writer = writer
+                    });
+                }
+                TypeWriters[type] = writers;
+            }
+            foreach (var writer in writers)
+            {
+                if (writer.Write(o, sb))
+                    hasLine = true;
+            }
+            if (hasLine) sb.AppendLine();
+        }
+
+        private static Dictionary<Type, List<DiaTypePropertyWriter>> TypeWriters = new Dictionary<Type, List<DiaTypePropertyWriter>>();
+        internal static string[] SymTagEnumValues;
+
+        class DiaTypePropertyWriter
+        {
+            public static object[] EmptyParams = new object[0];
+            public bool Unimplemented;
+            public System.Reflection.MethodInfo MI;
+            public string PropertyName;
+            public Func<string, object, StringBuilder, bool> Writer;
+            internal bool Write(object instance, StringBuilder sb)
+            {
+                if (Unimplemented)
+                    return false;
+                try
+                {
+                    object value = MI.Invoke(instance, EmptyParams);
+                    return Writer(PropertyName, value, sb);
+                }
+                catch (System.Reflection.TargetInvocationException tie)
+                {
+                    if (tie.InnerException is NotImplementedException)
+                    {
+                        this.Unimplemented = true;
+                    }
+                    return false;
+                }
+                catch (System.Runtime.InteropServices.COMException)
+                {
+                    return false;
+                }
+            }
+        }
+
+        private static bool WriteDiaValueAny(string propertyName, object propertyValue, StringBuilder sb)
+        {
+            if (null == propertyValue) return false;
+            if (System.Runtime.InteropServices.Marshal.IsComObject(propertyValue)) return false;
+            sb.Append(propertyName);
+            sb.Append(": ");
+            sb.Append(Convert.ToString(propertyValue));
+            sb.AppendLine();
+            return true;
+        }
+
+        private static bool WriteDiaValueSymTag(string propertyName, object propertyValueObj, StringBuilder sb)
+        {
+            uint tag = (uint)propertyValueObj;
+            sb.Append(propertyName);
+            sb.Append(": ");
+            sb.AppendLine(SymTagEnumValues[tag]);
+            return true;
+        }
+
+        private static bool WriteDiaValueBool(string propertyName, object propertyValueObj, StringBuilder sb)
+        {
+            bool propertyValue = (bool)propertyValueObj;
+            if (false == propertyValue) return false;
+            sb.Append(propertyName);
+            sb.Append(": ");
+            sb.Append(propertyValue);
+            sb.AppendLine();
+            return true;
+        }
+
+        private static bool WriteDiaValueUInt32(string propertyName, object propertyValueObj, StringBuilder sb)
+        {
+            uint propertyValue = (uint)propertyValueObj;
+            if (0 == propertyValue) return false;
+            sb.Append(propertyName);
+            sb.Append(": ");
+            sb.Append(propertyValue);
+            sb.AppendLine();
+            return true;
+        }
+
+        private static bool WriteDiaValueUInt64(string propertyName, object propertyValueObj, StringBuilder sb)
+        {
+            UInt64 propertyValue = (UInt64)propertyValueObj;
+            if (0 == propertyValue) return false;
+            sb.Append(propertyName);
+            sb.Append(": ");
+            sb.Append(propertyValue);
+            sb.AppendLine();
+            return true;
+        }
+
+        private static bool WriteDiaValueString(string propertyName, object propertyValueObj, StringBuilder sb)
+        {
+            string propertyValue = (string)propertyValueObj;
+            if (String.IsNullOrEmpty(propertyValue)) return false;
+            sb.Append(propertyName);
+            sb.Append(": ");
+            sb.AppendLine(propertyValue);
+            return true;
         }
     }
 
@@ -3076,431 +3315,5 @@ namespace MainNs
                 rtb.AppendLine(prefix + lines[i], c);
             }
         }
-    }
-}
-
-namespace dia2
-{
-    using System;
-    using System.Runtime.InteropServices;
-    using System.Runtime.InteropServices.ComTypes;
-
-    [ComImport]
-    [Guid("e6756135-1e65-4d17-8576-610761398c3c")]
-    class DiaDataSource
-    {
-
-    }
-
-    [ComImport]
-    [Guid("79F1BB5F-B66E-48e5-B6A9-1545C323CA3D")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    interface IDiaDataSource
-    {
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_lastError();
-        void loadDataFromPdb(string path);
-        void loadAndValidateDataFromPdb(string path, ref Guid pcsig70, UInt32 sig, UInt32 age);
-        void loadDataForExe(string executable, string searchPath, [MarshalAs(UnmanagedType.IUnknown)] object pCallback);
-        void loadDataFromIStream(IStream pIStream);
-        IDiaSession openSession();
-        // loadDataFromCodeViewInfo
-        // loadDataFromMiscInfo
-    }
-
-    [ComImport]
-    [Guid("2F609EE1-D1C8-4E24-8288-3326BADCD211")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    interface IDiaSession
-    {
-        UInt64 get_loadAddress();
-        void put_loadAddress(UInt64 value);
-        IDiaSymbol get_globalScope();
-        IDiaEnumTables getEnumTables();
-        void getSymbolsByAddr();
-        void findChildren();
-        void findChildrenEx();
-        void findChildrenExByAddr();
-        void findChildrenExByVA();
-        void findChildrenExByRVA();
-        void findSymbolByAddr();
-        void findSymbolByRVA();
-        void findSymbolByVA();
-        void findSymbolByToken();
-        void symsAreEquiv();
-        void symbolById();
-        void findSymbolByRVAEx();
-        void findSymbolByVAEx();
-        void findFile();
-        void findFileById();
-        void findLines();
-        void findLinesByAddr();
-        void findLinesByRVA();
-        void findLinesByVA();
-        void findLinesByLinenum();
-        object /*IDiaEnumInjectedSources*/ findInjectedSource(string srcFile);
-        object /*IDiaEnumDebugStreams*/ getEnumDebugStreams();
-        void findInlineFramesByAddr();
-        void findInlineFramesByRVA();
-        void findInlineFramesByVA();
-        void findInlineeLines();
-        void findInlineeLinesByAddr();
-        void findInlineeLinesByRVA();
-        void findInlineeLinesByVA();
-        void findInlineeLinesByLinenum();
-        void findInlineesByName();
-        void findAcceleratorInlineeLinesByLinenum();
-        void findSymbolsForAcceleratorPointerTag();
-        void findSymbolsByRVAForAcceleratorPointerTag();
-        void findAcceleratorInlineesByName();
-        void addressForVA();
-        void addressForRVA();
-        void findILOffsetsByAddr();
-        void findILOffsetsByRVA();
-        void findILOffsetsByVA();
-        void findInputAssemblyFiles();
-        void findInputAssembly();
-        void findInputAssemblyById();
-        void getFuncMDTokenMapSize();
-        void getFuncMDTokenMap();
-        void getTypeMDTokenMapSize();
-        void getTypeMDTokenMap();
-        void getNumberOfFunctionFragments_VA();
-        void getNumberOfFunctionFragments_RVA();
-        void getFunctionFragments_VA();
-        void getFunctionFragments_RVA();
-        object /*IDiaEnumSymbols*/ getExports();
-        object /*IDiaEnumSymbols*/ getHeapAllocationSites();
-        void findInputAssemblyFile();
-    }
-
-    [ComImport]
-    [Guid("cb787b2f-bd6c-4635-ba52-933126bd2dcd")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    interface IDiaSymbol
-    {
-        UInt32 get_symIndexId();
-        UInt32 get_symTag();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_name();
-        IDiaSymbol get_lexicalParent();
-        IDiaSymbol get_classParent();
-        IDiaSymbol get_type();
-        UInt32 get_dataKind();
-        UInt32 get_locationType();
-        UInt32 get_addressSection();
-        UInt32 get_addressOffset();
-        UInt32 get_relativeVirtualAddress();
-        UInt64 get_virtualAddress();
-        UInt32 get_registerId();
-        Int32 get_offset();
-        UInt64 get_length();
-        UInt32 get_slot();
-        bool get_volatileType();
-        bool get_constType();
-        bool get_unalignedType();
-        UInt32 get_access();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_libraryName();
-        UInt32 get_platform();
-        UInt32 get_language();
-        bool get_editAndContinueEnabled();
-        UInt32 get_frontEndMajor();
-        UInt32 get_frontEndMinor();
-        UInt32 get_frontEndBuild();
-        UInt32 get_backEndMajor();
-        UInt32 get_backEndMinor();
-        UInt32 get_backEndBuild();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_sourceFileName();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_unused();
-        UInt32 get_thunkOrdinal();
-        Int32 get_thisAdjust();
-        UInt32 get_virtualBaseOffset();
-        bool get_virtual();
-        bool get_intro();
-        bool get_pure();
-        UInt32 get_callingConvention();
-        [return: MarshalAs(UnmanagedType.AsAny)] // VARIANT
-        object get_value();
-        UInt32 get_baseType();
-        UInt32 get_token();
-        UInt32 get_timeStamp();
-        Guid get_guid();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_symbolsFileName();
-        bool get_reference();
-        UInt32 get_count();
-        UInt32 get_bitPosition();
-        IDiaSymbol get_arrayIndexType();
-        bool get_packed();
-        bool get_constructor();
-        bool get_overloadedOperator();
-        bool get_nested();
-        bool get_hasNestedTypes();
-        bool get_hasAssignmentOperator();
-        bool get_hasCastOperator();
-        bool get_scoped();
-        bool get_virtualBaseClass();
-        bool get_indirectVirtualBaseClass();
-        Int32 get_virtualBasePointerOffset();
-        IDiaSymbol get_virtualTableShape();
-        UInt32 get_lexicalParentId();
-        UInt32 get_classParentId();
-        UInt32 get_typeId();
-        UInt32 get_arrayIndexTypeId();
-        UInt32 get_virtualTableShapeId();
-        bool get_code();
-        bool get_function();
-        bool get_managed();
-        bool get_msil();
-        UInt32 get_virtualBaseDispIndex();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_undecoratedName();
-        UInt32 get_age();
-        UInt32 get_signature();
-        bool get_compilerGenerated();
-        bool get_addressTaken();
-        UInt32 get_rank();
-        IDiaSymbol get_lowerBound();
-        IDiaSymbol get_upperBound();
-        UInt32 get_lowerBoundId();
-        UInt32 get_upperBoundId();
-        void get_dataBytes(UInt32 cbData, out UInt32 pcbData, out byte[] pbData);
-        void findChildren();
-        void findChildrenEx();
-        void findChildrenExByAddr();
-        void findChildrenExByVA();
-        void findChildrenExByRVA();
-        UInt32 get_targetSection();
-        UInt32 get_targetOffset();
-        UInt32 get_targetRelativeVirtualAddress();
-        UInt64 get_targetVirtualAddress();
-        UInt32 get_machineType();
-        UInt32 get_oemId();
-        UInt32 get_oemSymbolId();
-        void get_types();
-        void get_typeIds();
-        IDiaSymbol get_objectPointerType();
-        UInt32 get_udtKind();
-        void get_undecoratedNameEx();
-        bool get_noReturn();
-        bool get_customCallingConvention();
-        bool get_noInline();
-        bool get_optimizedCodeDebugInfo();
-        bool get_notReached();
-        bool get_interruptReturn();
-        bool get_farReturn();
-        bool get_isStatic();
-        bool get_hasDebugInfo();
-        bool get_isLTCG();
-        bool get_isDataAligned();
-        bool get_hasSecurityChecks();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_compilerName();
-        bool get_hasAlloca();
-        bool get_hasSetJump();
-        bool get_hasLongJump();
-        bool get_hasInlAsm();
-        bool get_hasEH();
-        bool get_hasSEH();
-        bool get_hasEHa();
-        bool get_isNaked();
-        bool get_isAggregated();
-        bool get_isSplitted();
-        IDiaSymbol get_container();
-        bool get_inlSpec();
-        bool get_noStackOrdering();
-        IDiaSymbol get_virtualBaseTableType();
-        bool get_hasManagedCode();
-        bool get_isHotpatchable();
-        bool get_isCVTCIL();
-        bool get_isMSILNetmodule();
-        bool get_isCTypes();
-        bool get_isStripped();
-        UInt32 get_frontEndQFE();
-        UInt32 get_backEndQFE();
-        bool get_wasInlined();
-        bool get_strictGSCheck();
-        bool get_isCxxReturnUdt();
-        bool get_isConstructorVirtualBase();
-        bool get_RValueReference();
-        IDiaSymbol get_unmodifiedType();
-        bool get_framePointerPresent();
-        bool get_isSafeBuffers();
-        bool get_intrinsic();
-        bool get_sealed();
-        bool get_hfaFloat();
-        bool get_hfaDouble();
-        UInt32 get_liveRangeStartAddressSection();
-        UInt32 get_liveRangeStartAddressOffset();
-        UInt32 get_liveRangeStartRelativeVirtualAddress();
-        UInt32 get_countLiveRanges();
-        UInt64 get_liveRangeLength();
-        UInt32 get_offsetInUdt();
-        UInt32 get_paramBasePointerRegisterId();
-        UInt32 get_localBasePointerRegisterId();
-        bool get_isLocationControlFlowDependent();
-        UInt32 get_stride();
-        UInt32 get_numberOfRows();
-        UInt32 get_numberOfColumns();
-        bool get_isMatrixRowMajor();
-        void get_numericProperties();
-        void get_modifierValues();
-        bool get_isReturnValue();
-        bool get_isOptimizedAway();
-        UInt32 get_builtInKind();
-        UInt32 get_registerType();
-        UInt32 get_baseDataSlot();
-        UInt32 get_baseDataOffset();
-        UInt32 get_textureSlot();
-        UInt32 get_samplerSlot();
-        UInt32 get_uavSlot();
-        UInt32 get_sizeInUdt();
-        UInt32 get_memorySpaceKind();
-        UInt32 get_unmodifiedTypeId();
-        UInt32 get_subTypeId();
-        IDiaSymbol get_subType();
-        UInt32 get_numberOfModifiers();
-        UInt32 get_numberOfRegisterIndices();
-        bool get_isHLSLData();
-        bool get_isPointerToDataMember();
-        bool get_isPointerToMemberFunction();
-        bool get_isSingleInheritance();
-        bool get_isMultipleInheritance();
-        bool get_isVirtualInheritance();
-        bool get_restrictedType();
-        bool get_isPointerBasedOnSymbolValue();
-        IDiaSymbol get_baseSymbol();
-        UInt32 get_baseSymbolId();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_objectFileName();
-        bool get_isAcceleratorGroupSharedLocal();
-        bool get_isAcceleratorPointerTagLiveRange();
-        bool get_isAcceleratorStubFunction();
-        UInt32 get_numberOfAcceleratorPointerTags();
-        bool get_isSdl();
-        bool get_isWinRTPointer();
-        bool get_isRefUdt();
-        bool get_isValueUdt();
-        bool get_isInterfaceUdt();
-        void findInlineFramesByAddr();
-        void findInlineFramesByRVA();
-        void findInlineFramesByVA();
-        void findInlineeLines();
-        void findInlineeLinesByAddr();
-        void findInlineeLinesByRVA();
-        void findInlineeLinesByVA();
-        void findSymbolsForAcceleratorPointerTag();
-        void findSymbolsByRVAForAcceleratorPointerTag();
-        void get_acceleratorPointerTags();
-        void getSrcLineOnTypeDefn();
-        bool get_isPGO();
-        bool get_hasValidPGOCounts();
-        bool get_isOptimizedForSpeed();
-        UInt32 get_PGOEntryCount();
-        UInt32 get_PGOEdgeCount();
-        UInt64 get_PGODynamicInstructionCount();
-        UInt32 get_staticSize();
-        UInt32 get_finalLiveStaticSize();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_phaseName();
-        bool get_hasControlFlowCheck();
-        bool get_constantExport();
-        bool get_dataExport();
-        bool get_privateExport();
-        bool get_noNameExport();
-        bool get_exportHasExplicitlyAssignedOrdinal();
-        bool get_exportIsForwarder();
-        UInt32 get_ordinal();
-        UInt32 get_frameSize();
-        UInt32 get_exceptionHandlerAddressSection();
-        UInt32 get_exceptionHandlerAddressOffset();
-        UInt32 get_exceptionHandlerRelativeVirtualAddress();
-        UInt64 get_exceptionHandlerVirtualAddress();
-        void findInputAssemblyFile();
-        UInt32 get_characteristics();
-        IDiaSymbol get_coffGroup();
-        UInt32 get_bindID();
-        UInt32 get_bindSpace();
-        UInt32 get_bindSlot();
-    }
-
-    [ComImport]
-    [Guid("C65C2B0A-1150-4d7a-AFCC-E05BF3DEE81E")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    interface IDiaEnumTables
-    {
-        void get__NewEnum();
-        UInt32 get_Count();
-        IDiaTable Item(object index);
-        UInt32 Next(UInt32 count,
-            [MarshalAs(UnmanagedType.LPArray, ArraySubType = UnmanagedType.Interface, SizeParamIndex = 2)]
-            ref IDiaTable[] tables, out UInt32 fetched);
-        void Skip(UInt32 count);
-        void Reset();
-        IDiaEnumTables Clone();
-    }
-
-    [ComImport]
-    [Guid("00000100-0000-0000-C000-000000000046")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    interface IEnumUnknown
-    {
-        UInt32 Next(UInt32 count,
-            [MarshalAs(UnmanagedType.LPArray, ArraySubType = UnmanagedType.IUnknown, SizeParamIndex = 2)]
-            ref object[] tables, out UInt32 fetched);
-        void Skip(UInt32 count);
-        void Reset();
-        IEnumUnknown Clone();
-    }
-
-    [ComImport]
-    [Guid("4A59FB77-ABAC-469b-A30B-9ECC85BFEF14")]
-    [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
-    interface IDiaTable : IEnumUnknown
-    {
-        [return: MarshalAs(UnmanagedType.IUnknown)]
-        object get__NewEnum();
-        [return: MarshalAs(UnmanagedType.BStr)]
-        string get_name();
-        Int32 get_Count();
-        [return: MarshalAs(UnmanagedType.IUnknown)]
-        object Item(UInt32 index);
-    }
-}
-
-namespace D3DCompiler
-{
-    using System;
-    using System.Runtime.InteropServices;
-    using System.Runtime.InteropServices.ComTypes;
-
-    public struct D3D_SHADER_MACRO
-    {
-        [MarshalAs(UnmanagedType.LPStr)] string Name;
-        [MarshalAs(UnmanagedType.LPStr)] string Definition;
-    }
-
-    internal static class D3DCompiler
-    {
-        [DllImport("d3dcompiler_47.dll", CallingConvention = CallingConvention.Winapi, SetLastError = false, CharSet = CharSet.Ansi, ExactSpelling = true)]
-        public extern static Int32 D3DCompile(
-            [MarshalAs(UnmanagedType.LPStr)] string srcData, int srcDataSize,
-            [MarshalAs(UnmanagedType.LPStr)] string sourceName,
-            [MarshalAs(UnmanagedType.LPArray)] D3D_SHADER_MACRO[] defines,
-            int pInclude,
-            [MarshalAs(UnmanagedType.LPStr)] string entryPoint,
-            [MarshalAs(UnmanagedType.LPStr)] string target,
-            UInt32 Flags1,
-            UInt32 Flags2,
-            out IDxcBlob code, out IDxcBlob errorMsgs);
-
-        [DllImport("d3dcompiler_47.dll", CallingConvention = CallingConvention.Winapi, SetLastError = false, CharSet = CharSet.Ansi, ExactSpelling = true)]
-        public extern static Int32 D3DDisassemble(
-            IntPtr ptr, uint ptrSize, uint flags,
-            [MarshalAs(UnmanagedType.LPStr)] string szComments,
-            out IDxcBlob disassembly);
     }
 }
